+++ src/Terrabuild.Configuration.Tests/Workspace.fs
@@ -57,6 +57,7 @@ let parseWorkspace() =
 
         { WorkspaceFile.Workspace = { Id = "d7528db2-83e0-4164-8c8e-1e0d6d6357ca" |> Some
                                       Ignores = [ "**/node_modules" ] |> Set |> Some
+                                      Deny = [ ".git" ] |> Set |> Some
                                       Version = None
                                       Engine = "docker" |> Some
                                       Configuration = "local" |> Some 
@@ -121,6 +122,7 @@ let parseWorkspace2() =
         { WorkspaceFile.Workspace =
             { Id = None
               Ignores = None
+              Deny = None
               Version = None
               Engine = None
               Configuration = None