+++ src/Terrabuild.Tests/Core/Configuration.fs
@@ -23,7 +23,7 @@ let ``Extension script path must stay inside workspace``() =
     Directory.CreateDirectory(workspace) |> ignore
 
     try
-        let loader = Extensions.lazyLoadScript workspace "@custom" (Some "../outside.fss")
+        let loader = Extensions.lazyLoadScript workspace [ ".git" ] "@custom" (Some "../outside.fss")
         (fun () -> loader.Value |> ignore)
         |> should (throwWithMessage $"Script '../outside.fss' is outside workspace '{workspace}'") typeof<TerrabuildException>
     finally
@@ -32,6 +32,6 @@ let ``Extension script path must stay inside workspace``() =
 [<Test>]
 let ``HTTP extension script URL is rejected``() =
     let workspace = Path.GetTempPath()
-    let loader = Extensions.lazyLoadScript workspace "@custom" (Some "http://example.com/extension.fss")
+    let loader = Extensions.lazyLoadScript workspace [ ".git" ] "@custom" (Some "http://example.com/extension.fss")
     (fun () -> loader.Value |> ignore)
     |> should (throwWithMessage "Only HTTPS script URLs are allowed for extension '@custom'") typeof<TerrabuildException>